Move window management sources to 'wm/'.
