+++ src/framework/wm/events.cc
@@ -1,4 +1,4 @@
-#include "framework/events.h"
+#include "framework/wm/events.h"
 
 #include <algorithm>
