+++ src/framework/wm/window.cc
@@ -1,6 +1,6 @@
-#include "framework/window.h"
+#include "framework/wm/window.h"
 #include "framework/common.h"
-#include "framework/events.h"
+#include "framework/wm/events.h"
 
 /* -------------------------------------------------------------------------- */
